Read description...
The code that searches for the entry works and the actual form works as well. The code that puts the entry into the database is where the error is. I have tried to change the SQL and that hasn't worked. I have also been comparing this code to the practice database code.